Maintenance: refactor ExternalLink component
