--- src/components/shared/Extensions/ExternalLink/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ExternalLink.tsx';
