--- src/components/shared/Extensions/ExternalLink/types.ts
@@ -1,14 +1,0 @@
-import { type ReactNode } from 'react';
-import { type MessageDescriptor } from 'react-intl';
-
-import { type SimpleMessageValues } from '~types/index.ts';
-
-export interface ExternalLinkProps {
-  children?: ReactNode;
-  href: string;
-  text?: MessageDescriptor | string;
-  textValues?: SimpleMessageValues;
-  className?: string;
-  title?: string;
-  download?: string | boolean;
-}
