+++ src/components/shared/ExternalLink/types.ts
@@ -1,0 +1,5 @@
+import { type HTMLProps } from 'react';
+
+export interface ExternalLinkProps extends HTMLProps<HTMLAnchorElement> {
+  hasHover?: boolean;
+}
