+++ src/components/shared/TransactionLink/types.ts
@@ -1,0 +1,6 @@
+import { type ExternalLinkProps } from '~shared/ExternalLink/types.ts';
+
+export interface TransactionLinkProps extends ExternalLinkProps {
+  /** Transaction hash */
+  hash: string;
+}
